.github/scripts/requirements.txt
requests
tiktoken
pyyaml
lxml
